Refactor DateUtils method calls
